move poly bench to arkworks
